## 2.0.4 - Features:   - Now a track selection persists when editing or adding.   - Metronome stops when navigating to import/edit forms from player, which helps with a lot of concurrency issues. - Bugfixes:   - Fix defect with fetching BPM in edit forms.     - Was not using the most recently update title/artist when editing.   - Fix metronome disabled color (made it gray).     - Always enable stop, in case metronome is still playing the silent track after skipping from previous track.   - Fix notes display in track player.     - No longer needs to expand, since drawer handles that.   - Fixed concurrency issues with writing click tracks.   - Fixed several issues with selected track and binding to audio sources (metronome). - Tech Debt:   - Rewrite concept of "added item".  Instead of a null, it creates the item instance first.     - This lets forms bind to the actual field values, and saving just performs agnostic insert/update into the DB.   - Also remove all optional setlistTrack references when launching edit/import forms.     - Set list track is now required for any related pages/forms launch, so optional params are removed from all constructors, argument type classes, etc.

+++ lib/ui/pages/metronome_page.dart
@@ -4,8 +4,8 @@ import 'package:in_the_pocket/ui/controls/metronome.dart';
 import 'package:provider/provider.dart';
 
 const double DEFAULT_BPM = 120.0;
-const double DEFAULT_TIME_SIGNATURE_TOP = 4.0;
-const double DEFAULT_TIME_SIGNATURE_BOTTOM = 4.0;
+const int DEFAULT_TIME_SIGNATURE_TOP = 4;
+const int DEFAULT_TIME_SIGNATURE_BOTTOM = 4;
 
 class MetronomePage extends StatefulWidget {
   const MetronomePage({Key? key}) : super(key: key);
